Refactor: 상품 조회 URI 변경. Admin -> 일반
